refactor: move the v2 contacts entitlement guard behind the license-check boundary (ENG-2295)
