+++ apps/web/modules/api/v2/management/contact-attribute-keys/[contactAttributeKeyId]/route.ts
@@ -13,7 +13,7 @@ import {
   ZContactAttributeKeyUpdateSchema,
 } from "@/modules/api/v2/management/contact-attribute-keys/[contactAttributeKeyId]/types/contact-attribute-keys";
 import { ApiErrorResponseV2 } from "@/modules/api/v2/types/api-error";
-import { checkContactsEnabledApiV2 } from "@/modules/ee/contacts/lib/contacts-entitlement";
+import { checkContactsEnabledApiV2 } from "@/modules/ee/license-check/lib/contacts-api-guard";
 import { hasPermission } from "@/modules/organization/settings/api-keys/lib/utils";
 
 export const GET = async (
