+++ apps/web/modules/ee/contacts/lib/contacts-entitlement.test.ts
@@ -9,8 +9,7 @@ vi.mock("@/modules/ee/license-check/lib/utils", () => ({
   getIsContactsEnabled: mocks.getIsContactsEnabled,
 }));
 
-const { CONTACTS_NOT_ENABLED_MESSAGE, checkContactsEnabledApiV2, ensureContactsEnabled } =
-  await import("./contacts-entitlement");
+const { CONTACTS_NOT_ENABLED_MESSAGE, ensureContactsEnabled } = await import("./contacts-entitlement");
 
 describe("ensureContactsEnabled", () => {
   beforeEach(() => {
@@ -29,28 +28,5 @@ describe("ensureContactsEnabled", () => {
     mocks.getIsContactsEnabled.mockResolvedValue(true);
 
     await expect(ensureContactsEnabled("org1")).resolves.toBeUndefined();
-  });
-});
-
-describe("checkContactsEnabledApiV2", () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  test("returns a forbidden error object when the entitlement is missing", async () => {
-    mocks.getIsContactsEnabled.mockResolvedValue(false);
-
-    const error = await checkContactsEnabledApiV2("org1");
-
-    expect(error).toEqual({
-      type: "forbidden",
-      details: [{ field: "contacts", issue: "Contacts feature is not enabled for this organization" }],
-    });
-  });
-
-  test("returns null when the entitlement is present", async () => {
-    mocks.getIsContactsEnabled.mockResolvedValue(true);
-
-    await expect(checkContactsEnabledApiV2("org1")).resolves.toBeNull();
   });
 });
